[Trivial] Fix all 36 mypy --strict errors and make pre-commit blocking
- Fix type annotation issues across 13 src/ files (no behavioral changes)
- Install types-PyYAML for YAML stub types
- Make mypy check blocking in pre-commit hook (was warning-only)
- Expand mypy coverage to all staged src/ files

Error categories fixed:
- no-any-return: explicit type annotations on dynamic returns
- return-value: PermissionResult vs tuple mismatch in permission_checker
- redundant-cast: removed unnecessary casts in runner.py, world.py
- unused-ignore: removed stale type: ignore comments
- attr-defined: added __all__ exports in executor.py
- arg-type: guard None values in loader.py
- assignment: fixed variable reuse bug in runner.py
- typeddict-item: bridge BalanceInfo types in checkpoint.py
- type-arg: add generic params to dict in run_manager.py

+++ src/dashboard/api/routes/metrics.py
@@ -130,7 +130,7 @@ async def get_efficiency_leaderboard(limit: int = 10) -> LeaderboardResponse:
     def get_success_rate(a: Any) -> float:
         if a.action_count == 0:
             return 0.0
-        return a.action_successes / a.action_count
+        return float(a.action_successes / a.action_count)
 
     sorted_agents = sorted(agents, key=get_success_rate, reverse=True)[:limit]
